+++ examples/set_get.php
@@ -1,24 +1,18 @@
 <?php
 
 use seregazhuk\React\Memcached\Factory;
-use seregazhuk\React\Memcached\Client;
 
 require '../vendor/autoload.php';
 
 $loop = React\EventLoop\Factory::create();
-$factory = new Factory($loop);
+$client = Factory::createClient($loop);
+$client->set('example', 'Hello world');
 
-$factory
-    ->createClient('localhost:11211')
-    ->then(function (Client $client) {
-        $client->set('example', 'Hello world');
-
-        $client->get('example')->then(function ($data) {
-            echo $data . PHP_EOL; // Hello world
-        });
-
-        // Close the connection when all requests are resolved
-        $client->end();
+$client->get('example')->then(function ($data) {
+    echo $data . PHP_EOL; // Hello world
 });
+
+// Close the connection when all requests are resolved
+$client->end();
 
 $loop->run();