+++ src/Exception/ConnectionClosedException.php
@@ -6,7 +6,7 @@ use Throwable;
 
 class ConnectionClosedException extends Exception
 {
-    public function __construct($message = "", $code = 0, Throwable $previous = null)
+    public function __construct($message = '', $code = 0, Throwable $previous = null)
     {
         parent::__construct('Connection closed', $code, $previous);
     }