+++ src/Factory.php
@@ -3,8 +3,6 @@
 namespace seregazhuk\React\Memcached;
 
 use React\EventLoop\LoopInterface;
-use React\Promise\PromiseInterface;
-use React\Socket\ConnectionInterface;
 use React\Socket\Connector;
 use seregazhuk\React\Memcached\Protocol\Parser;
 use seregazhuk\React\Memcached\Protocol\Response\Factory as ResponseFactory;
@@ -13,40 +11,31 @@ use seregazhuk\React\Memcached\Protocol\Request\Factory as RequestFactory;
 class Factory
 {
     /**
-     * @var Connector
-     */
-    private $connector;
-
-    /**
+     * Creates a memcached client
      * @param LoopInterface $loop
+     * @param string $address
+     * @return Client
      */
-    public function __construct(LoopInterface $loop)
+    public static function createClient(LoopInterface $loop, $address = 'localhost:11211')
     {
-        $this->connector = new Connector($loop);
-    }
-
-    /**
-     * Creates a memcached client connected to a given connection string
-     *
-     * @param string $address Memcached server URI to connect to
-     * @return PromiseInterface resolves with Client or rejects with \RuntimeException
-     */
-    public function createClient($address = 'localhost:11211')
-    {
-        $promise = $this->connector
-            ->connect($address)
-            ->then(function (ConnectionInterface $stream) {
-                return new Client($stream, $this->createProtocolParser());
-            });
-
-        return $promise;
+        return new Client(self::createConnection($loop, $address),  self::createProtocolParser());
     }
 
     /**
      * @return Parser
      */
-    private function createProtocolParser()
+    protected static function createProtocolParser()
     {
         return new Parser(new RequestFactory(), new ResponseFactory());
+    }
+
+    /**
+     * @param LoopInterface $loop
+     * @param $address
+     * @return Connection
+     */
+    protected static function createConnection(LoopInterface $loop, $address)
+    {
+        return new Connection($address, new Connector($loop));
     }
 }