+++ src/Protocol/Response/ReadResponse.php
@@ -14,9 +14,9 @@ class ReadResponse extends Response
         $regExp = '/VALUE \w+ \d+ \d+' . Parser::COMMAND_SEPARATOR . '(.*)' . Parser::COMMAND_SEPARATOR . 'END/';
         preg_match($regExp, $this->data, $match);
 
-        $value = isset($match[1]) ? $match[1] : null;
+        $value = isset($match[1]) ? trim($match[1]) : null;
 
-        if(is_null($value)) {
+        if(null === $value) {
             $this->fail();
         }