+++ tests/Unit/ClientTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace seregazhuk\React\Memcached\tests;
+namespace seregazhuk\React\Memcached\tests\Unit;
 
 use Mockery;
 use Mockery\MockInterface;
 use React\Stream\DuplexStreamInterface;
 use seregazhuk\React\Memcached\Client;
+use seregazhuk\React\Memcached\Connection;
 use seregazhuk\React\Memcached\Exception\ConnectionClosedException;
 use seregazhuk\React\Memcached\Exception\Exception;
 use seregazhuk\React\Memcached\Exception\WrongCommandException;
@@ -14,11 +15,7 @@ use seregazhuk\React\Memcached\Protocol\Request\Factory as RequestFactory;
 use seregazhuk\React\Memcached\Protocol\Response\Factory as ResponseFactory;
 
 class ClientTest extends PromiseTestCase
-{    /**
- * @var DuplexStreamInterface|MockInterface
- */
-    protected $stream;
-
+{
     /**
      * @var Parser|MockInterface
      */
@@ -29,11 +26,16 @@ class ClientTest extends PromiseTestCase
      */
     protected $client;
 
+    /**
+     * @var Connection|MockInterface
+     */
+    protected $connection;
+
     protected function setUp()
     {
-        $this->stream = Mockery::mock(DuplexStreamInterface::class)->shouldReceive('on')->getMock();
         $this->parser = Mockery::mock(Parser::class, [new RequestFactory(), new ResponseFactory()])->makePartial();
-        $this->client = new Client($this->stream, $this->parser);
+        $this->connection = Mockery::mock(Connection::class)->shouldReceive('on')->getMock();
+        $this->client = new Client($this->connection, $this->parser);
 
         parent::setUp();
     }
@@ -42,7 +44,7 @@ class ClientTest extends PromiseTestCase
     public function it_sends_data_to_the_connection()
     {
         $this->parser->shouldReceive('makeRequest')->andReturn("version\n\r");
-        $this->stream->shouldReceive('write')->with("version\n\r")->once();
+        $this->connection->shouldReceive('write')->with("version\n\r")->once();
         $this->client->version();
     }
 
@@ -57,7 +59,7 @@ class ClientTest extends PromiseTestCase
     public function it_resolves_a_promise_with_data_from_response()
     {
         $this->parser->shouldReceive('makeRequest')->once();
-        $this->stream->shouldReceive('write')->once();
+        $this->connection->shouldReceive('write')->once();
         $promise = $this->client->version();
 
         $this->client->resolveRequests(['12345']);
@@ -76,8 +78,8 @@ class ClientTest extends PromiseTestCase
     public function it_rejects_pending_request_when_closing()
     {
         $this->parser->shouldReceive('makeRequest')->once();
-        $this->stream->shouldReceive('write')->once();
-        $this->stream->shouldReceive('close')->once();
+        $this->connection->shouldReceive('write')->once();
+        $this->connection->shouldReceive('close')->once();
         $promise = $this->client->version();
 
         $this->client->close();
@@ -89,7 +91,7 @@ class ClientTest extends PromiseTestCase
     public function it_rejects_all_new_requests_when_closed()
     {
         $this->parser->shouldNotReceive('makeRequest');
-        $this->stream->shouldReceive('close')->once();
+        $this->connection->shouldReceive('close')->once();
 
         $this->client->close();
         $promise = $this->client->version();
@@ -100,7 +102,7 @@ class ClientTest extends PromiseTestCase
     public function it_rejects_all_new_requests_when_ending()
     {
         $this->parser->shouldNotReceive('makeRequest');
-        $this->stream->shouldReceive('close')->once();
+        $this->connection->shouldReceive('close')->once();
 
         $this->client->end();
         $promise = $this->client->version();
@@ -110,7 +112,7 @@ class ClientTest extends PromiseTestCase
 	/** @test */
 	public function it_emits_close_event_when_closing()
 	{
-		$this->stream->shouldReceive('close')->once();
+		$this->connection->shouldReceive('close')->once();
 
 		$this->client->on('close', $this->expectCallableOnce());
 		$this->client->end();